feat: update shared db
